fix(desktop): split secret_store and marker helpers to satisfy file-size ratchet
secret_store.rs (1307 lines pre-change) and app_state.rs (1055) are
inherited-oversize files the CI ratchet forbids growing. Move the
interprocess lock to secret_store_lock.rs, the debug file backend to
secret_store_file.rs (each with their tests), and the migration-marker
helpers to app_state_keyring.rs where the marker naming already lives.
No behavior change.
