+++ desktop/src-tauri/src/secret_store.rs
@@ -36,7 +36,6 @@
 //! divergent-behavior trap.
 
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::Mutex;
 
 /// Result of probing the keyring before a migration: distinguishes "reachable
@@ -58,283 +57,18 @@ pub enum KeyringProbe {
 /// as a JSON map under this name within the service.
 const BLOB_KEY: &str = "secrets";
 
-// ── Interprocess advisory lock ─────────────────────────────────────────────
-//
-// Two concurrent Buzz processes (e.g. the signed DMG build and an unsigned dev
-// build via `just staging`) share the same OS keychain blob because the
-// service name `"buzz-desktop"` is a constant — it does not key off the bundle
-// identifier. Each process holds its own in-memory cache, so without an
-// interprocess lock a warm-cache write in process A drops keys added by process
-// B between A's last cache-warming read and A's write.
-//
-// The fix: `mutate_blob` acquires an exclusive advisory file lock, then always
-// performs a fresh `read_blob_raw()` inside the lock, applies the mutation,
-// writes back, and releases. The cache is still updated after a successful
-// write, so same-process reads remain fast. The lock is file-based at a fixed
-// per-user path `/tmp/buzz-keychain-<uid>-<service>.lock` on Unix — a path
-// that is invariant to `$TMPDIR`/process environment, so both the GUI-launched
-// signed DMG and a terminal-launched dev build always take the same lock.
-
-/// Return the path of the advisory lockfile for `service`.
-///
-/// The path is `/tmp/buzz-keychain-<uid>-<service>.lock` on Unix — a
-/// deterministic per-user path that is invariant to `$TMPDIR`/process
-/// environment. Both a GUI-launched signed DMG (`launchd`, env-stripped) and a
-/// terminal-launched dev build resolve `/tmp` to the same inode, so they
-/// contend on the same lockfile and achieve mutual exclusion.
-///
-/// On Windows the same name used for the kernel mutex is derived from the
-/// lockfile path, so the service-keyed uniqueness is preserved.
-fn blob_lockfile_path(service: &str) -> PathBuf {
-    #[cfg(unix)]
-    {
-        // Use the real UID so distinct users get distinct lockfiles.
-        // SAFETY: getuid() is always safe on Unix — it never fails.
-        let uid = unsafe { libc::getuid() };
-        PathBuf::from(format!("/tmp/buzz-keychain-{uid}-{service}.lock"))
-    }
-    #[cfg(not(unix))]
-    {
-        // Windows: no lockfile used (named mutex instead); this path is only
-        // used to derive the mutex name and for test assertions.
-        std::env::temp_dir().join(format!("buzz-keychain-{service}.lock"))
-    }
-}
-
-/// Acquire an exclusive advisory file lock for the blob identified by `service`.
-///
-/// Opens (or creates) the lockfile and blocks until the lock is acquired.
-/// Returns the open `File`; the lock is released when the file is dropped.
-///
-/// On non-Unix/non-Windows platforms this is a no-op that returns a stub.
+#[path = "secret_store_lock.rs"]
+mod lock;
 #[cfg(feature = "system-keyring")]
-fn acquire_blob_lock(service: &str) -> Result<BlobLockGuard, String> {
-    let path = blob_lockfile_path(service);
-    BlobLockGuard::acquire(&path)
-}
+use lock::acquire_blob_lock;
 
-/// RAII guard that holds an exclusive advisory file lock.
-///
-/// On Unix, implemented via `flock(2)` on a lockfile in the system temp dir.
-/// On Windows, implemented via a named kernel mutex (cross-process, no file I/O
-/// needed). The Windows mutex handle is released on drop.
-#[cfg(feature = "system-keyring")]
-struct BlobLockGuard {
-    /// The open lockfile. Never read — held purely for RAII: closing the fd
-    /// releases the `flock(LOCK_EX)` on Unix.
-    #[cfg(unix)]
-    #[allow(dead_code)]
-    file: std::fs::File,
-    #[cfg(windows)]
-    mutex_handle: windows_sys::Win32::Foundation::HANDLE,
-}
-
-#[cfg(feature = "system-keyring")]
-impl BlobLockGuard {
-    fn acquire(path: &std::path::Path) -> Result<Self, String> {
-        #[cfg(unix)]
-        {
-            let file = std::fs::OpenOptions::new()
-                .create(true)
-                .truncate(false)
-                .write(true)
-                .open(path)
-                .map_err(|e| format!("blob lock open {}: {e}", path.display()))?;
-            use std::os::unix::io::AsRawFd;
-            // LOCK_EX blocks until the lock is acquired (no LOCK_NB).
-            let ret = unsafe { libc::flock(file.as_raw_fd(), libc::LOCK_EX) };
-            if ret != 0 {
-                let err = std::io::Error::last_os_error();
-                return Err(format!("blob lock flock: {err}"));
-            }
-            return Ok(BlobLockGuard { file });
-        }
-
-        #[cfg(windows)]
-        {
-            // Named kernel mutexes are cross-process on Windows — no lockfile
-            // needed. Derive a unique mutex name from the lockfile path so
-            // distinct services get distinct mutexes.
-            let name_str = format!(
-                "Local\\BuzzKeychain-{}",
-                path.file_stem()
-                    .and_then(|s| s.to_str())
-                    .unwrap_or("default")
-            );
-            // Encode as null-terminated UTF-16.
-            let name_wide: Vec<u16> = name_str
-                .encode_utf16()
-                .chain(std::iter::once(0u16))
-                .collect();
-            use windows_sys::Win32::Foundation::WAIT_OBJECT_0;
-            use windows_sys::Win32::Security::SECURITY_ATTRIBUTES;
-            use windows_sys::Win32::System::Threading::{
-                CreateMutexW, WaitForSingleObject, INFINITE,
-            };
-            // CreateMutexW: lpMutexAttributes = null (default security),
-            // bInitialOwner = FALSE (0), lpName = our mutex name.
-            let handle = unsafe {
-                CreateMutexW(
-                    std::ptr::null::<SECURITY_ATTRIBUTES>(),
-                    0,
-                    name_wide.as_ptr(),
-                )
-            };
-            // HANDLE = *mut c_void; null means creation failed.
-            if handle.is_null() {
-                let err = std::io::Error::last_os_error();
-                return Err(format!("blob lock CreateMutexW: {err}"));
-            }
-            let wait_result = unsafe { WaitForSingleObject(handle, INFINITE) };
-            if wait_result != WAIT_OBJECT_0 {
-                // Also accept WAIT_ABANDONED (0x80) — previous holder crashed;
-                // the mutex is still acquired and we own it.
-                if wait_result != windows_sys::Win32::Foundation::WAIT_ABANDONED {
-                    let err = std::io::Error::last_os_error();
-                    unsafe { windows_sys::Win32::Foundation::CloseHandle(handle) };
-                    return Err(format!(
-                        "blob lock WaitForSingleObject: {wait_result} / {err}"
-                    ));
-                }
-            }
-            return Ok(BlobLockGuard {
-                mutex_handle: handle,
-            });
-        }
-
-        // Fallback for exotic platforms: no-op lock (only Unix/Windows ship).
-        #[allow(unreachable_code)]
-        Err("blob lock: unsupported platform".to_string())
-    }
-}
-
-#[cfg(feature = "system-keyring")]
-impl Drop for BlobLockGuard {
-    fn drop(&mut self) {
-        #[cfg(unix)]
-        {
-            // Dropping `self.file` closes the fd, which releases flock on Unix.
-            // Nothing explicit needed.
-        }
-        #[cfg(windows)]
-        {
-            unsafe {
-                windows_sys::Win32::System::Threading::ReleaseMutex(self.mutex_handle);
-                windows_sys::Win32::Foundation::CloseHandle(self.mutex_handle);
-            }
-        }
-    }
-}
-
-// ── End interprocess advisory lock ────────────────────────────────────────
-
-/// Read the file backend's blob. `Ok(None)` = no file yet (fresh store —
-/// deliberately no fallback to the old dev keychain item).
+#[path = "secret_store_file.rs"]
+mod file_backend;
+#[cfg(debug_assertions)]
+pub use file_backend::init_file_backend_dir;
+use file_backend::{backend_for, SecretBackend};
 #[cfg(all(debug_assertions, feature = "system-keyring"))]
-fn read_blob_raw_file(path: &std::path::Path) -> Result<Option<Vec<u8>>, String> {
-    match std::fs::read(path) {
-        Ok(bytes) => Ok(Some(bytes)),
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(None),
-        Err(e) => Err(format!("secrets file read {}: {e}", path.display())),
-    }
-}
-
-/// Atomically replace the file backend's blob: write a `0o600` sibling tmp
-/// file, fsync, rename over the final path. A crash mid-write can never leave
-/// a truncated secrets file.
-#[cfg(all(debug_assertions, feature = "system-keyring"))]
-fn write_blob_raw_file(path: &std::path::Path, bytes: &[u8]) -> Result<(), String> {
-    use std::io::Write;
-
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent)
-            .map_err(|e| format!("secrets dir create {}: {e}", parent.display()))?;
-    }
-    let file_name = path
-        .file_name()
-        .and_then(|n| n.to_str())
-        .unwrap_or("secrets.json");
-    let tmp = path.with_file_name(format!("{file_name}.tmp"));
-
-    let mut opts = std::fs::OpenOptions::new();
-    opts.write(true).create(true).truncate(true);
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::OpenOptionsExt;
-        opts.mode(0o600);
-    }
-    let mut file = opts
-        .open(&tmp)
-        .map_err(|e| format!("secrets tmp open {}: {e}", tmp.display()))?;
-    file.write_all(bytes)
-        .map_err(|e| format!("secrets tmp write: {e}"))?;
-    file.sync_all()
-        .map_err(|e| format!("secrets tmp fsync: {e}"))?;
-    drop(file);
-    std::fs::rename(&tmp, path).map_err(|e| format!("secrets file rename: {e}"))
-}
-
-/// Where a [`SecretStore`]'s blob physically lives. The `File` variant is
-/// debug-only so release binaries are keyring-only by construction.
-enum SecretBackend {
-    Keyring,
-    /// Blob at this exact path (`secrets.<service>.json` in the app-data dir).
-    #[cfg(debug_assertions)]
-    File(PathBuf),
-}
-
-/// App-data dir for the debug file backend, recorded once at boot.
-#[cfg(debug_assertions)]
-static FILE_BACKEND_DIR: std::sync::OnceLock<PathBuf> = std::sync::OnceLock::new();
-
-/// Record the app-data dir the debug file backend stores secrets under.
-/// Must run before the first [`SecretStore::shared`] call — `lib.rs` setup
-/// does, ahead of `run_boot_reset`. Later calls are no-ops.
-#[cfg(debug_assertions)]
-pub fn init_file_backend_dir(dir: &std::path::Path) {
-    let _ = FILE_BACKEND_DIR.set(dir.to_path_buf());
-}
-
-/// Pure backend decision for debug builds: keyring when the user opted back
-/// in via `BUZZ_DEV_USE_KEYCHAIN=1` or the file dir was never initialized,
-/// otherwise the per-service secrets file (namespaced because `just dev` and
-/// a main-checkout standalone share one app-data dir with different services).
-#[cfg(debug_assertions)]
-fn select_backend(
-    use_keychain_env: Option<&str>,
-    file_dir: Option<&std::path::Path>,
-    service: &str,
-) -> SecretBackend {
-    if use_keychain_env == Some("1") {
-        return SecretBackend::Keyring;
-    }
-    match file_dir {
-        Some(dir) => SecretBackend::File(dir.join(format!("secrets.{service}.json"))),
-        None => SecretBackend::Keyring,
-    }
-}
-
-#[cfg(debug_assertions)]
-fn backend_for(service: &str) -> SecretBackend {
-    let env = std::env::var("BUZZ_DEV_USE_KEYCHAIN").ok();
-    let backend = select_backend(
-        env.as_deref(),
-        FILE_BACKEND_DIR.get().map(|p| p.as_path()),
-        service,
-    );
-    if matches!(backend, SecretBackend::Keyring) && env.as_deref() != Some("1") {
-        eprintln!(
-            "buzz-desktop: file backend dir not initialized; \
-             using OS keychain for service {service}"
-        );
-    }
-    backend
-}
-
-#[cfg(not(debug_assertions))]
-fn backend_for(_service: &str) -> SecretBackend {
-    SecretBackend::Keyring
-}
+use file_backend::{read_blob_raw_file, write_blob_raw_file};
 
 /// Secret storage addressed by service name. All secrets are stored in a
 /// single JSON blob (one OS prompt per process lifetime on the keyring
@@ -1234,64 +968,6 @@ mod tests {
         let _ = reader.delete("agent_b");
     }
 
-    #[test]
-    fn test_blob_lockfile_path_is_in_tmp_with_uid() {
-        // The lockfile must be at a deterministic per-user path under /tmp —
-        // invariant to $TMPDIR — so both a GUI-launched DMG (env-stripped by
-        // launchd) and a terminal-launched dev build resolve the same inode and
-        // achieve mutual exclusion.
-        let path = blob_lockfile_path("buzz-desktop");
-        #[cfg(unix)]
-        {
-            let uid = unsafe { libc::getuid() };
-            assert!(
-                path.starts_with("/tmp"),
-                "lockfile {path:?} must start with /tmp (not $TMPDIR)"
-            );
-            let name = path
-                .file_name()
-                .and_then(|n| n.to_str())
-                .unwrap_or_default();
-            assert!(
-                name.contains(&uid.to_string()),
-                "lockfile {path:?} must contain uid {uid}"
-            );
-            assert!(
-                name.contains("buzz-keychain"),
-                "lockfile name must contain 'buzz-keychain'"
-            );
-        }
-        #[cfg(not(unix))]
-        {
-            assert!(
-                path.file_name()
-                    .and_then(|n| n.to_str())
-                    .is_some_and(|n| n.contains("buzz-keychain")),
-                "lockfile name must contain 'buzz-keychain'"
-            );
-        }
-    }
-
-    #[test]
-    fn test_blob_lock_acquire_and_release() {
-        // Verify the advisory lock can be acquired and released without errors.
-        // This exercises the real flock/mutex path on the current platform.
-        let guard = acquire_blob_lock("buzz-test-lock-smoke");
-        assert!(
-            guard.is_ok(),
-            "advisory lock acquire must succeed: {:?}",
-            guard.err()
-        );
-        // Drop the guard — lock is released. A second acquire must succeed.
-        drop(guard);
-        let guard2 = acquire_blob_lock("buzz-test-lock-smoke");
-        assert!(
-            guard2.is_ok(),
-            "advisory lock re-acquire after release must succeed: {:?}",
-            guard2.err()
-        );
-    }
-
     #[ignore = "requires real OS keychain (run locally)"]
     #[test]
     fn mutate_blob_does_not_advance_cache_on_write_failure() {
@@ -1449,185 +1125,6 @@ mod tests {
 
         // Cleanup.
         let _ = store2.delete(key);
-    }
-
-    // ── Debug file backend (no OS keychain; CI-runnable) ───────────────────
-
-    #[cfg(debug_assertions)]
-    mod file_backend {
-        use super::*;
-
-        // Test-only constructor: file backend at an explicit path, bypassing
-        // the process-global FILE_BACKEND_DIR.
-        impl SecretStore {
-            fn file_at(service: &str, path: std::path::PathBuf) -> Self {
-                SecretStore {
-                    service: service.to_string(),
-                    backend: SecretBackend::File(path),
-                    cache: Mutex::new(None),
-                }
-            }
-        }
-
-        fn tmp_store(service: &str) -> (tempfile::TempDir, SecretStore) {
-            let dir = tempfile::tempdir().unwrap();
-            let path = dir.path().join(format!("secrets.{service}.json"));
-            let store = SecretStore::file_at(service, path);
-            (dir, store)
-        }
-
-        #[test]
-        fn select_backend_uses_service_namespaced_file_when_dir_set() {
-            let dir = std::path::Path::new("/tmp/buzz-test-data");
-            match select_backend(None, Some(dir), "buzz-desktop-dev.slug") {
-                SecretBackend::File(p) => {
-                    assert_eq!(p, dir.join("secrets.buzz-desktop-dev.slug.json"))
-                }
-                SecretBackend::Keyring => panic!("expected file backend"),
-            }
-        }
-
-        #[test]
-        fn select_backend_env_escape_hatch_forces_keyring() {
-            let dir = std::path::Path::new("/tmp/buzz-test-data");
-            assert!(matches!(
-                select_backend(Some("1"), Some(dir), "buzz-desktop-dev"),
-                SecretBackend::Keyring
-            ));
-            // Any value other than "1" does not opt back into the keychain.
-            assert!(matches!(
-                select_backend(Some("0"), Some(dir), "buzz-desktop-dev"),
-                SecretBackend::File(_)
-            ));
-        }
-
-        #[test]
-        fn select_backend_without_dir_falls_back_to_keyring() {
-            assert!(matches!(
-                select_backend(None, None, "buzz-desktop-dev"),
-                SecretBackend::Keyring
-            ));
-        }
-
-        #[test]
-        fn is_file_backed_reflects_backend() {
-            let (_dir, store) = tmp_store("buzz-test-file-backed");
-            assert!(store.is_file_backed());
-            assert!(!SecretStore::keyring("buzz-test-file-backed").is_file_backed());
-        }
-
-        #[test]
-        fn file_roundtrip_store_load_delete() {
-            let (_dir, store) = tmp_store("buzz-test-file-roundtrip");
-            store.store("identity", "nsec1aaa").unwrap();
-            store.store("agent:abc", "nsec1bbb").unwrap();
-            assert_eq!(
-                store.load("identity").unwrap(),
-                Some("nsec1aaa".to_string())
-            );
-            assert_eq!(
-                store.load("agent:abc").unwrap(),
-                Some("nsec1bbb".to_string())
-            );
-            store.delete("agent:abc").unwrap();
-            assert_eq!(store.load("agent:abc").unwrap(), None);
-            assert_eq!(
-                store.load("identity").unwrap(),
-                Some("nsec1aaa".to_string())
-            );
-        }
-
-        #[test]
-        fn missing_key_probe_and_load_never_consult_legacy_keychain() {
-            // A fresh (empty) file store must report reachable-but-empty and
-            // Ok(None) without falling through to the legacy keychain
-            // migration paths — file mode never touches the OS keychain.
-            let (_dir, store) = tmp_store("buzz-test-file-no-legacy");
-            assert_eq!(store.probe("identity"), KeyringProbe::ReachableButEmpty);
-            assert_eq!(store.load("identity").unwrap(), None);
-            // Same when a blob exists but the key is absent.
-            store.store("other", "v").unwrap();
-            assert_eq!(store.probe("identity"), KeyringProbe::ReachableButEmpty);
-            assert_eq!(store.load("identity").unwrap(), None);
-            assert_eq!(store.probe("other"), KeyringProbe::Present);
-        }
-
-        #[cfg(unix)]
-        #[test]
-        fn file_is_created_0o600() {
-            use std::os::unix::fs::PermissionsExt;
-            let (dir, store) = tmp_store("buzz-test-file-perms");
-            store.store("identity", "nsec1aaa").unwrap();
-            let path = dir.path().join("secrets.buzz-test-file-perms.json");
-            let mode = std::fs::metadata(&path).unwrap().permissions().mode();
-            assert_eq!(mode & 0o777, 0o600, "secrets file must be 0o600");
-        }
-
-        #[test]
-        fn write_is_atomic_leaves_no_tmp_residue() {
-            let (dir, store) = tmp_store("buzz-test-file-atomic");
-            store.store("identity", "nsec1aaa").unwrap();
-            store.store("agent:abc", "nsec1bbb").unwrap();
-            let names: Vec<String> = std::fs::read_dir(dir.path())
-                .unwrap()
-                .map(|e| e.unwrap().file_name().to_string_lossy().into_owned())
-                .collect();
-            assert_eq!(
-                names,
-                vec!["secrets.buzz-test-file-atomic.json".to_string()],
-                "only the final secrets file may remain: {names:?}"
-            );
-        }
-
-        #[test]
-        fn corrupt_file_fails_closed_and_is_preserved() {
-            let (dir, store) = tmp_store("buzz-test-file-corrupt");
-            let path = dir.path().join("secrets.buzz-test-file-corrupt.json");
-            std::fs::write(&path, b"not json").unwrap();
-            assert!(store.load("identity").is_err());
-            assert_eq!(store.probe("identity"), KeyringProbe::Unreachable);
-            // A store() must not clobber the corrupt file — the fresh read
-            // inside mutate_blob fails first.
-            assert!(store.store("identity", "nsec1aaa").is_err());
-            assert_eq!(std::fs::read(&path).unwrap(), b"not json");
-        }
-
-        #[test]
-        fn two_stores_same_path_observe_each_others_writes() {
-            // CI-runnable port of the cross-process stale-cache race test:
-            // two instances (= two processes with separate caches) on one
-            // file must never drop each other's keys.
-            let dir = tempfile::tempdir().unwrap();
-            let svc = "buzz-test-file-race";
-            let path = dir.path().join(format!("secrets.{svc}.json"));
-            let store_a = SecretStore::file_at(svc, path.clone());
-            store_a.store("k1", "v1").unwrap(); // warms A's cache
-            let store_b = SecretStore::file_at(svc, path.clone());
-            store_b.store("k2", "v2").unwrap();
-            store_a.store("k3", "v3").unwrap(); // must re-read, not drop k2
-            let reader = SecretStore::file_at(svc, path);
-            for (k, v) in [("k1", "v1"), ("k2", "v2"), ("k3", "v3")] {
-                assert_eq!(
-                    reader.load(k).unwrap(),
-                    Some(v.to_string()),
-                    "{k} must survive"
-                );
-            }
-        }
-
-        #[test]
-        fn delete_all_removes_file_and_verifies_wiped() {
-            let (dir, store) = tmp_store("buzz-test-file-wipe");
-            store.store("identity", "nsec1aaa").unwrap();
-            assert!(!store.verify_fully_wiped());
-            store.delete_all_with_legacy_cleanup().unwrap();
-            let path = dir.path().join("secrets.buzz-test-file-wipe.json");
-            assert!(!path.exists(), "secrets file must be deleted");
-            assert!(store.verify_fully_wiped());
-            // Idempotent on an already-absent file.
-            store.delete_all_with_legacy_cleanup().unwrap();
-            assert_eq!(store.load("identity").unwrap(), None);
-        }
     }
 
     #[ignore = "requires real OS keychain (run locally)"]
