+++ desktop/src-tauri/src/secret_store_lock.rs
@@ -1,0 +1,232 @@
+//! Interprocess advisory lock for the [`super::SecretStore`] blob.
+//!
+//! Two concurrent Buzz processes (e.g. the signed DMG build and an unsigned dev
+//! build via `just staging`) share the same OS keychain blob because the
+//! service name `"buzz-desktop"` is a constant — it does not key off the bundle
+//! identifier. Each process holds its own in-memory cache, so without an
+//! interprocess lock a warm-cache write in process A drops keys added by process
+//! B between A's last cache-warming read and A's write.
+//!
+//! The fix: `mutate_blob` acquires an exclusive advisory file lock, then always
+//! performs a fresh `read_blob_raw()` inside the lock, applies the mutation,
+//! writes back, and releases. The cache is still updated after a successful
+//! write, so same-process reads remain fast. The lock is file-based at a fixed
+//! per-user path `/tmp/buzz-keychain-<uid>-<service>.lock` on Unix — a path
+//! that is invariant to `$TMPDIR`/process environment, so both the GUI-launched
+//! signed DMG and a terminal-launched dev build always take the same lock.
+
+use std::path::PathBuf;
+
+/// Return the path of the advisory lockfile for `service`.
+///
+/// The path is `/tmp/buzz-keychain-<uid>-<service>.lock` on Unix — a
+/// deterministic per-user path that is invariant to `$TMPDIR`/process
+/// environment. Both a GUI-launched signed DMG (`launchd`, env-stripped) and a
+/// terminal-launched dev build resolve `/tmp` to the same inode, so they
+/// contend on the same lockfile and achieve mutual exclusion.
+///
+/// On Windows the same name used for the kernel mutex is derived from the
+/// lockfile path, so the service-keyed uniqueness is preserved.
+fn blob_lockfile_path(service: &str) -> PathBuf {
+    #[cfg(unix)]
+    {
+        // Use the real UID so distinct users get distinct lockfiles.
+        // SAFETY: getuid() is always safe on Unix — it never fails.
+        let uid = unsafe { libc::getuid() };
+        PathBuf::from(format!("/tmp/buzz-keychain-{uid}-{service}.lock"))
+    }
+    #[cfg(not(unix))]
+    {
+        // Windows: no lockfile used (named mutex instead); this path is only
+        // used to derive the mutex name and for test assertions.
+        std::env::temp_dir().join(format!("buzz-keychain-{service}.lock"))
+    }
+}
+
+/// Acquire an exclusive advisory file lock for the blob identified by `service`.
+///
+/// Opens (or creates) the lockfile and blocks until the lock is acquired.
+/// Returns the open `File`; the lock is released when the file is dropped.
+///
+/// On non-Unix/non-Windows platforms this is a no-op that returns a stub.
+#[cfg(feature = "system-keyring")]
+pub(super) fn acquire_blob_lock(service: &str) -> Result<BlobLockGuard, String> {
+    let path = blob_lockfile_path(service);
+    BlobLockGuard::acquire(&path)
+}
+
+/// RAII guard that holds an exclusive advisory file lock.
+///
+/// On Unix, implemented via `flock(2)` on a lockfile in the system temp dir.
+/// On Windows, implemented via a named kernel mutex (cross-process, no file I/O
+/// needed). The Windows mutex handle is released on drop.
+#[cfg(feature = "system-keyring")]
+pub(super) struct BlobLockGuard {
+    /// The open lockfile. Never read — held purely for RAII: closing the fd
+    /// releases the `flock(LOCK_EX)` on Unix.
+    #[cfg(unix)]
+    #[allow(dead_code)]
+    file: std::fs::File,
+    #[cfg(windows)]
+    mutex_handle: windows_sys::Win32::Foundation::HANDLE,
+}
+
+#[cfg(feature = "system-keyring")]
+impl BlobLockGuard {
+    fn acquire(path: &std::path::Path) -> Result<Self, String> {
+        #[cfg(unix)]
+        {
+            let file = std::fs::OpenOptions::new()
+                .create(true)
+                .truncate(false)
+                .write(true)
+                .open(path)
+                .map_err(|e| format!("blob lock open {}: {e}", path.display()))?;
+            use std::os::unix::io::AsRawFd;
+            // LOCK_EX blocks until the lock is acquired (no LOCK_NB).
+            let ret = unsafe { libc::flock(file.as_raw_fd(), libc::LOCK_EX) };
+            if ret != 0 {
+                let err = std::io::Error::last_os_error();
+                return Err(format!("blob lock flock: {err}"));
+            }
+            return Ok(BlobLockGuard { file });
+        }
+
+        #[cfg(windows)]
+        {
+            // Named kernel mutexes are cross-process on Windows — no lockfile
+            // needed. Derive a unique mutex name from the lockfile path so
+            // distinct services get distinct mutexes.
+            let name_str = format!(
+                "Local\\BuzzKeychain-{}",
+                path.file_stem()
+                    .and_then(|s| s.to_str())
+                    .unwrap_or("default")
+            );
+            // Encode as null-terminated UTF-16.
+            let name_wide: Vec<u16> = name_str
+                .encode_utf16()
+                .chain(std::iter::once(0u16))
+                .collect();
+            use windows_sys::Win32::Foundation::WAIT_OBJECT_0;
+            use windows_sys::Win32::Security::SECURITY_ATTRIBUTES;
+            use windows_sys::Win32::System::Threading::{
+                CreateMutexW, WaitForSingleObject, INFINITE,
+            };
+            // CreateMutexW: lpMutexAttributes = null (default security),
+            // bInitialOwner = FALSE (0), lpName = our mutex name.
+            let handle = unsafe {
+                CreateMutexW(
+                    std::ptr::null::<SECURITY_ATTRIBUTES>(),
+                    0,
+                    name_wide.as_ptr(),
+                )
+            };
+            // HANDLE = *mut c_void; null means creation failed.
+            if handle.is_null() {
+                let err = std::io::Error::last_os_error();
+                return Err(format!("blob lock CreateMutexW: {err}"));
+            }
+            let wait_result = unsafe { WaitForSingleObject(handle, INFINITE) };
+            if wait_result != WAIT_OBJECT_0 {
+                // Also accept WAIT_ABANDONED (0x80) — previous holder crashed;
+                // the mutex is still acquired and we own it.
+                if wait_result != windows_sys::Win32::Foundation::WAIT_ABANDONED {
+                    let err = std::io::Error::last_os_error();
+                    unsafe { windows_sys::Win32::Foundation::CloseHandle(handle) };
+                    return Err(format!(
+                        "blob lock WaitForSingleObject: {wait_result} / {err}"
+                    ));
+                }
+            }
+            return Ok(BlobLockGuard {
+                mutex_handle: handle,
+            });
+        }
+
+        // Fallback for exotic platforms: no-op lock (only Unix/Windows ship).
+        #[allow(unreachable_code)]
+        Err("blob lock: unsupported platform".to_string())
+    }
+}
+
+#[cfg(feature = "system-keyring")]
+impl Drop for BlobLockGuard {
+    fn drop(&mut self) {
+        #[cfg(unix)]
+        {
+            // Dropping `self.file` closes the fd, which releases flock on Unix.
+            // Nothing explicit needed.
+        }
+        #[cfg(windows)]
+        {
+            unsafe {
+                windows_sys::Win32::System::Threading::ReleaseMutex(self.mutex_handle);
+                windows_sys::Win32::Foundation::CloseHandle(self.mutex_handle);
+            }
+        }
+    }
+}
+
+#[cfg(all(test, feature = "system-keyring"))]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_blob_lockfile_path_is_in_tmp_with_uid() {
+        // The lockfile must be at a deterministic per-user path under /tmp —
+        // invariant to $TMPDIR — so both a GUI-launched DMG (env-stripped by
+        // launchd) and a terminal-launched dev build resolve the same inode and
+        // achieve mutual exclusion.
+        let path = blob_lockfile_path("buzz-desktop");
+        #[cfg(unix)]
+        {
+            let uid = unsafe { libc::getuid() };
+            assert!(
+                path.starts_with("/tmp"),
+                "lockfile {path:?} must start with /tmp (not $TMPDIR)"
+            );
+            let name = path
+                .file_name()
+                .and_then(|n| n.to_str())
+                .unwrap_or_default();
+            assert!(
+                name.contains(&uid.to_string()),
+                "lockfile {path:?} must contain uid {uid}"
+            );
+            assert!(
+                name.contains("buzz-keychain"),
+                "lockfile name must contain 'buzz-keychain'"
+            );
+        }
+        #[cfg(not(unix))]
+        {
+            assert!(
+                path.file_name()
+                    .and_then(|n| n.to_str())
+                    .is_some_and(|n| n.contains("buzz-keychain")),
+                "lockfile name must contain 'buzz-keychain'"
+            );
+        }
+    }
+
+    #[test]
+    fn test_blob_lock_acquire_and_release() {
+        // Verify the advisory lock can be acquired and released without errors.
+        // This exercises the real flock/mutex path on the current platform.
+        let guard = acquire_blob_lock("buzz-test-lock-smoke");
+        assert!(
+            guard.is_ok(),
+            "advisory lock acquire must succeed: {:?}",
+            guard.err()
+        );
+        // Drop the guard — lock is released. A second acquire must succeed.
+        drop(guard);
+        let guard2 = acquire_blob_lock("buzz-test-lock-smoke");
+        assert!(
+            guard2.is_ok(),
+            "advisory lock re-acquire after release must succeed: {:?}",
+            guard2.err()
+        );
+    }
+}
